requirements.txt
amqp==2.6.1
asgiref==3.2.10
billiard==3.6.3.0
celery==4.4.7
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
colorlog==4.6.2
coreapi==2.3.3
coreschema==0.0.4
Django==3.1.1
django-celery-beat==2.1.0
django-celery-results==2.0.0
django-filter==2.4.0
django-rest-swagger==2.2.0
django-timezone-field==4.1.1
djangorestframework==3.12.2
dnspython==1.16.0
drf-yasg==1.20.0
eventlet==0.30.0
flower==0.9.7
greenlet==0.4.17
humanize==3.2.0
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.2
kombu==4.6.11
Markdown==3.3.3
MarkupSafe==1.1.1
openapi-codec==1.3.2
packaging==20.7
Pillow==7.2.0
prometheus-client==0.8.0
prompt-toolkit==3.0.8
psycopg2-binary==2.8.6
pyparsing==2.4.7
python-crontab==2.5.1
python-dateutil==2.8.1
python-decouple==3.3
pytz==2020.1
PyYAML==5.3.1
redis==3.5.3
requests==2.24.0
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
simplejson==3.17.2
six==1.15.0
sqlparse==0.4.1
tornado==6.1
uritemplate==3.0.1
urllib3==1.25.10
vine==1.3.0
wcwidth==0.2.5
